start generalization of front-end Home page
