+++ docker-compose.local.yml
@@ -92,6 +92,7 @@ services:
     container_name: local_frontend
     volumes:
       - "./frontend/src:/app/src"
+      - "./frontend/public/stats:/app/public/stats"
     build:
       context: .
       dockerfile: frontend.development.dockerfile
